+++ pkg/sql/schemachanger/scbuild/internal/scbuildstmt/type_helpers.go
@@ -17,6 +17,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/catconstants"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/catid"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/sqlerrors"
 	"github.com/cockroachdb/cockroach/pkg/util/log/eventpb"
 )
 
@@ -163,6 +164,109 @@ func setSchemaForTypeDesc(
 		NewDescriptorName: newName.FQString(),
 		DescriptorType:    descriptorType,
 	})
+}
+
+// renameForTypeDesc implements RENAME TO for user-defined types that have an
+// associated array type (enums, composites, domains).
+func renameForTypeDesc(
+	b BuildCtx,
+	tn *tree.TypeName,
+	typeElem scpb.Element,
+	typeID catid.DescID,
+	arrayTypeID catid.DescID,
+	newName string,
+) {
+	typeElts := b.QueryByID(typeID)
+	currentNS := typeElts.FilterNamespace().NotToAbsent().MustGetOneElement()
+
+	// Unlike table rename (which no-ops on self-rename), Postgres treats type
+	// rename to the same name as an error.
+	if currentNS.Name == newName {
+		typeName := tree.MakeTableNameFromPrefix(
+			b.NamePrefix(currentNS), tree.Name(newName),
+		)
+		panic(sqlerrors.NewTypeAlreadyExistsError(typeName.String()))
+	}
+
+	checkTypeNameConflicts(b, newName, currentNS)
+
+	newNS := &scpb.Namespace{
+		DatabaseID:   currentNS.DatabaseID,
+		SchemaID:     currentNS.SchemaID,
+		DescriptorID: currentNS.DescriptorID,
+		Name:         newName,
+	}
+	b.Drop(currentNS)
+	b.Add(newNS)
+
+	// Rename the implicit array type.
+	arrayElts := b.QueryByID(arrayTypeID)
+	arrayNS := arrayElts.FilterNamespace().NotToAbsent().MustGetOneElement()
+
+	newArrayName := findFreeNameInSchema(b, b.NamePrefix(typeElem), "_"+newName)
+	newArrayNS := &scpb.Namespace{
+		DatabaseID:   arrayNS.DatabaseID,
+		SchemaID:     arrayNS.SchemaID,
+		DescriptorID: arrayNS.DescriptorID,
+		Name:         newArrayName,
+	}
+	b.Drop(arrayNS)
+	b.Add(newArrayNS)
+
+	b.LogEventForExistingPayload(newNS, &eventpb.RenameType{
+		TypeName:    tn.FQString(),
+		NewTypeName: newName,
+	})
+}
+
+// checkTypeNameConflicts checks if the new type name conflicts with an
+// existing object in the same schema.
+func checkTypeNameConflicts(b BuildCtx, newName string, currentNS *scpb.Namespace) {
+	tn := tree.MakeTableNameFromPrefix(
+		b.NamePrefix(currentNS),
+		tree.Name(newName),
+	)
+	ers := b.ResolveRelation(tn.ToUnresolvedObjectName(),
+		ResolveParams{
+			IsExistenceOptional: true,
+			WithOffline:         true,
+			ResolveTypes:        true,
+		})
+	if ers == nil || ers.IsEmpty() {
+		return
+	}
+
+	existingNS := ers.FilterNamespace().NotToAbsent().MustGetZeroOrOneElement()
+	if existingNS == nil {
+		if !ers.FilterNamespace().ToAbsent().IsEmpty() {
+			panic(pgerror.Newf(pgcode.ObjectNotInPrerequisiteState,
+				"object %q being dropped, try again later", newName))
+		}
+		return
+	}
+
+	if !ers.FilterCompositeType().IsEmpty() || !ers.FilterEnumType().IsEmpty() || !ers.FilterDomainType().IsEmpty() {
+		panic(sqlerrors.NewTypeAlreadyExistsError(tn.String()))
+	}
+	panic(sqlerrors.NewRelationAlreadyExistsError(tn.String()))
+}
+
+// findFreeNameInSchema prepends underscores to candidateName until it doesn't
+// collide with any existing object in the given schema.
+func findFreeNameInSchema(b BuildCtx, prefix tree.ObjectNamePrefix, candidateName string) string {
+	for {
+		tn := tree.MakeTableNameFromPrefix(prefix, tree.Name(candidateName))
+		ers := b.ResolveRelation(tn.ToUnresolvedObjectName(),
+			ResolveParams{
+				IsExistenceOptional: true,
+				WithOffline:         true,
+				ResolveTypes:        true,
+			})
+		if ers == nil || ers.IsEmpty() {
+			return candidateName
+		}
+		candidateName = "_" + candidateName
+	}
 }
 
 // moveDescriptorToSchema drops the old Namespace and SchemaChild elements for
