+++ pkg/sql/schemachanger/scbuild/testdata/alter_domain
@@ -18,6 +18,18 @@ build skip=sql_dependencies
 ALTER DOMAIN defaultdb.d SET SCHEMA public
 ----
 
+build skip=sql_dependencies
+ALTER DOMAIN defaultdb.d RENAME TO d2
+----
+- [[Namespace:{DescID: 104, Name: d, ReferencedDescID: 100, IntValue: 101}, ABSENT], PUBLIC]
+  {databaseId: 100, descriptorId: 104, name: d, schemaId: 101}
+- [[Namespace:{DescID: 104, Name: d2, ReferencedDescID: 100, IntValue: 101}, PUBLIC], ABSENT]
+  {databaseId: 100, descriptorId: 104, name: d2, schemaId: 101}
+- [[Namespace:{DescID: 105, Name: _d, ReferencedDescID: 100, IntValue: 101}, ABSENT], PUBLIC]
+  {databaseId: 100, descriptorId: 105, name: _d, schemaId: 101}
+- [[Namespace:{DescID: 105, Name: _d2, ReferencedDescID: 100, IntValue: 101}, PUBLIC], ABSENT]
+  {databaseId: 100, descriptorId: 105, name: _d2, schemaId: 101}
+
 setup
 CREATE SCHEMA s;
 ----
