+++ pkg/sql/schemachanger/testdata/end_to_end/alter_domain_rename/alter_domain_rename.definition
@@ -1,0 +1,7 @@
+setup
+CREATE DOMAIN salmon AS INT DEFAULT 0;
+----
+
+test
+ALTER DOMAIN salmon RENAME TO trout;
+----
