+++ pkg/sql/schemachanger/testdata/end_to_end/alter_domain_rename/alter_domain_rename.explain
@@ -1,0 +1,48 @@
+/* setup */
+CREATE DOMAIN salmon AS INT DEFAULT 0;
+
+/* test */
+EXPLAIN (DDL) ALTER DOMAIN salmon RENAME TO trout;
+----
+Schema change plan for ALTER DOMAIN ‹defaultdb›.‹public›.‹salmon› RENAME TO ‹trout›;
+ ├── StatementPhase
+ │    └── Stage 1 of 1 in StatementPhase
+ │         ├── 2 elements transitioning toward PUBLIC
+ │         │    ├── ABSENT → PUBLIC Namespace:{DescID: 104 (salmon-trout+), Name: "trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         │    └── ABSENT → PUBLIC Namespace:{DescID: 105 (_salmon-_trout+), Name: "_trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         ├── 2 elements transitioning toward ABSENT
+ │         │    ├── PUBLIC → ABSENT Namespace:{DescID: 104 (salmon-trout+), Name: "salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         │    └── PUBLIC → ABSENT Namespace:{DescID: 105 (_salmon-_trout+), Name: "_salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         └── 6 Mutation operations
+ │              ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"salmon","SchemaID":101}}
+ │              ├── SetNameInDescriptor {"DescriptorID":104,"Name":"trout"}
+ │              ├── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"trout","SchemaID":101}}
+ │              ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"_salmon","SchemaID":101}}
+ │              ├── SetNameInDescriptor {"DescriptorID":105,"Name":"_trout"}
+ │              └── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"_trout","SchemaID":101}}
+ └── PreCommitPhase
+      ├── Stage 1 of 2 in PreCommitPhase
+      │    ├── 2 elements transitioning toward PUBLIC
+      │    │    ├── PUBLIC → ABSENT Namespace:{DescID: 104 (salmon-trout+), Name: "trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    │    └── PUBLIC → ABSENT Namespace:{DescID: 105 (_salmon-_trout+), Name: "_trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    ├── 2 elements transitioning toward ABSENT
+      │    │    ├── ABSENT → PUBLIC Namespace:{DescID: 104 (salmon-trout+), Name: "salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    │    └── ABSENT → PUBLIC Namespace:{DescID: 105 (_salmon-_trout+), Name: "_salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    └── 1 Mutation operation
+      │         └── UndoAllInTxnImmediateMutationOpSideEffects
+      └── Stage 2 of 2 in PreCommitPhase
+           ├── 2 elements transitioning toward PUBLIC
+           │    ├── ABSENT → PUBLIC Namespace:{DescID: 104 (salmon-trout+), Name: "trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           │    └── ABSENT → PUBLIC Namespace:{DescID: 105 (_salmon-_trout+), Name: "_trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           ├── 2 elements transitioning toward ABSENT
+           │    ├── PUBLIC → ABSENT Namespace:{DescID: 104 (salmon-trout+), Name: "salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           │    └── PUBLIC → ABSENT Namespace:{DescID: 105 (_salmon-_trout+), Name: "_salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           └── 8 Mutation operations
+                ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"salmon","SchemaID":101}}
+                ├── SetNameInDescriptor {"DescriptorID":104,"Name":"trout"}
+                ├── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"trout","SchemaID":101}}
+                ├── UpdateTTLScheduleMetadata {"NewName":"trout","TableID":104}
+                ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"_salmon","SchemaID":101}}
+                ├── SetNameInDescriptor {"DescriptorID":105,"Name":"_trout"}
+                ├── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"_trout","SchemaID":101}}
+                └── UpdateTTLScheduleMetadata {"NewName":"_trout","TableID":105}
