+++ pkg/sql/schemachanger/testdata/end_to_end/alter_domain_rename/alter_domain_rename.explain_shape
@@ -1,0 +1,8 @@
+/* setup */
+CREATE DOMAIN salmon AS INT DEFAULT 0;
+
+/* test */
+EXPLAIN (DDL, SHAPE) ALTER DOMAIN salmon RENAME TO trout;
+----
+Schema change plan for ALTER DOMAIN ‹defaultdb›.‹public›.‹salmon› RENAME TO ‹trout›;
+ └── execute 1 system table mutations transaction
