+++ pkg/sql/schemachanger/testdata/end_to_end/alter_domain_rename_collision/alter_domain_rename_collision.definition
@@ -1,0 +1,8 @@
+setup
+CREATE DOMAIN salmon AS INT DEFAULT 0;
+CREATE TYPE _trout AS ENUM ('rainbow');
+----
+
+test
+ALTER DOMAIN salmon RENAME TO trout;
+----
