+++ pkg/sql/schemachanger/testdata/end_to_end/alter_domain_rename_collision/alter_domain_rename_collision.explain
@@ -1,0 +1,49 @@
+/* setup */
+CREATE DOMAIN salmon AS INT DEFAULT 0;
+CREATE TYPE _trout AS ENUM ('rainbow');
+
+/* test */
+EXPLAIN (DDL) ALTER DOMAIN salmon RENAME TO trout;
+----
+Schema change plan for ALTER DOMAIN ‹defaultdb›.‹public›.‹salmon› RENAME TO ‹trout›;
+ ├── StatementPhase
+ │    └── Stage 1 of 1 in StatementPhase
+ │         ├── 2 elements transitioning toward PUBLIC
+ │         │    ├── ABSENT → PUBLIC Namespace:{DescID: 104 (salmon-trout+), Name: "trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         │    └── ABSENT → PUBLIC Namespace:{DescID: 105 (_salmon-___trout+), Name: "___trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         ├── 2 elements transitioning toward ABSENT
+ │         │    ├── PUBLIC → ABSENT Namespace:{DescID: 104 (salmon-trout+), Name: "salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         │    └── PUBLIC → ABSENT Namespace:{DescID: 105 (_salmon-___trout+), Name: "_salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+ │         └── 6 Mutation operations
+ │              ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"salmon","SchemaID":101}}
+ │              ├── SetNameInDescriptor {"DescriptorID":104,"Name":"trout"}
+ │              ├── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"trout","SchemaID":101}}
+ │              ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"_salmon","SchemaID":101}}
+ │              ├── SetNameInDescriptor {"DescriptorID":105,"Name":"___trout"}
+ │              └── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"___trout","SchemaID":101}}
+ └── PreCommitPhase
+      ├── Stage 1 of 2 in PreCommitPhase
+      │    ├── 2 elements transitioning toward PUBLIC
+      │    │    ├── PUBLIC → ABSENT Namespace:{DescID: 104 (salmon-trout+), Name: "trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    │    └── PUBLIC → ABSENT Namespace:{DescID: 105 (_salmon-___trout+), Name: "___trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    ├── 2 elements transitioning toward ABSENT
+      │    │    ├── ABSENT → PUBLIC Namespace:{DescID: 104 (salmon-trout+), Name: "salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    │    └── ABSENT → PUBLIC Namespace:{DescID: 105 (_salmon-___trout+), Name: "_salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+      │    └── 1 Mutation operation
+      │         └── UndoAllInTxnImmediateMutationOpSideEffects
+      └── Stage 2 of 2 in PreCommitPhase
+           ├── 2 elements transitioning toward PUBLIC
+           │    ├── ABSENT → PUBLIC Namespace:{DescID: 104 (salmon-trout+), Name: "trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           │    └── ABSENT → PUBLIC Namespace:{DescID: 105 (_salmon-___trout+), Name: "___trout", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           ├── 2 elements transitioning toward ABSENT
+           │    ├── PUBLIC → ABSENT Namespace:{DescID: 104 (salmon-trout+), Name: "salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           │    └── PUBLIC → ABSENT Namespace:{DescID: 105 (_salmon-___trout+), Name: "_salmon", ReferencedDescID: 100 (defaultdb), IntValue: 101}
+           └── 8 Mutation operations
+                ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"salmon","SchemaID":101}}
+                ├── SetNameInDescriptor {"DescriptorID":104,"Name":"trout"}
+                ├── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":104,"Name":"trout","SchemaID":101}}
+                ├── UpdateTTLScheduleMetadata {"NewName":"trout","TableID":104}
+                ├── DrainDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"_salmon","SchemaID":101}}
+                ├── SetNameInDescriptor {"DescriptorID":105,"Name":"___trout"}
+                ├── AddDescriptorName {"Namespace":{"DatabaseID":100,"DescriptorID":105,"Name":"___trout","SchemaID":101}}
+                └── UpdateTTLScheduleMetadata {"NewName":"___trout","TableID":105}
