+++ pkg/sql/schemachanger/testdata/end_to_end/alter_domain_rename_collision/alter_domain_rename_collision.side_effects
@@ -1,0 +1,95 @@
+/* setup */
+CREATE DOMAIN salmon AS INT DEFAULT 0;
+CREATE TYPE _trout AS ENUM ('rainbow');
+----
+...
++object {100 101 salmon} -> 104
++object {100 101 _salmon} -> 105
++object {100 101 _trout} -> 106
++object {100 101 __trout} -> 107
+
+/* test */
+ALTER DOMAIN salmon RENAME TO trout;
+----
+begin transaction #1
+# begin StatementPhase
+checking for feature: ALTER DOMAIN
+increment telemetry for sql.schema.alter_domain.rename
+write *eventpb.RenameType to event log:
+  newTypeName: trout
+  sql:
+    descriptorId: 104
+    statement: ALTER DOMAIN ‹defaultdb›.‹public›.‹salmon› RENAME TO ‹trout›
+    tag: ALTER DOMAIN
+    user: root
+  typeName: defaultdb.public.salmon
+## StatementPhase stage 1 of 1 with 6 MutationType ops
+delete object namespace entry {100 101 salmon} -> 104
+delete object namespace entry {100 101 _salmon} -> 105
+add object namespace entry {100 101 trout} -> 104
+add object namespace entry {100 101 ___trout} -> 105
+upsert descriptor #104
+  ...
+     kind: DOMAIN
+     modificationTime: {}
+  -  name: salmon
+  +  name: trout
+     parentId: 100
+     parentSchemaId: 101
+  ...
+         withGrantOption: "2"
+       version: 3
+  -  version: "1"
+  +  version: "2"
+upsert descriptor #105
+  ...
+     kind: ALIAS
+     modificationTime: {}
+  -  name: _salmon
+  +  name: ___trout
+     parentId: 100
+     parentSchemaId: 101
+  ...
+     referencingDescriptorIds:
+     - 104
+  -  version: "1"
+  +  version: "2"
+# end StatementPhase
+# begin PreCommitPhase
+## PreCommitPhase stage 1 of 2 with 1 MutationType op
+undo all catalog changes within txn #1
+persist all catalog changes to storage
+## PreCommitPhase stage 2 of 2 with 8 MutationType ops
+delete object namespace entry {100 101 salmon} -> 104
+delete object namespace entry {100 101 _salmon} -> 105
+add object namespace entry {100 101 trout} -> 104
+add object namespace entry {100 101 ___trout} -> 105
+upsert descriptor #104
+  ...
+     kind: DOMAIN
+     modificationTime: {}
+  -  name: salmon
+  +  name: trout
+     parentId: 100
+     parentSchemaId: 101
+  ...
+         withGrantOption: "2"
+       version: 3
+  -  version: "1"
+  +  version: "2"
+upsert descriptor #105
+  ...
+     kind: ALIAS
+     modificationTime: {}
+  -  name: _salmon
+  +  name: ___trout
+     parentId: 100
+     parentSchemaId: 101
+  ...
+     referencingDescriptorIds:
+     - 104
+  -  version: "1"
+  +  version: "2"
+persist all catalog changes to storage
+# end PreCommitPhase
+commit transaction #1
